chore: BackendAdminAPI 추가를 위한 준비 작업
