+++ packages/common/src/apis/client.ts
@@ -1,7 +1,8 @@
 import axios, { AxiosInstance, AxiosRequestConfig, AxiosResponse } from "axios";
 import * as R from "remeda";
 
-import CommonSchemas from "../schemas";
+import BackendAPISchemas from '../schemas/backendAPI';
+import { getCookie } from '../utils/cookie';
 
 const DEFAULT_ERROR_MESSAGE = "알 수 없는 문제가 발생했습니다, 잠시 후 다시 시도해주세요.";
 const DEFAULT_ERROR_RESPONSE = {
@@ -12,12 +13,12 @@ const DEFAULT_ERROR_RESPONSE = {
 export class BackendAPIClientError extends Error {
   readonly name = "BackendAPIClientError";
   readonly status: number;
-  readonly detail: CommonSchemas.ErrorResponseSchema;
+  readonly detail: BackendAPISchemas.ErrorResponseSchema;
   readonly originalError: unknown;
 
   constructor(error?: unknown) {
     let message: string = DEFAULT_ERROR_MESSAGE;
-    let detail: CommonSchemas.ErrorResponseSchema = DEFAULT_ERROR_RESPONSE;
+    let detail: BackendAPISchemas.ErrorResponseSchema = DEFAULT_ERROR_RESPONSE;
     let status = -1;
 
     if (axios.isAxiosError(error)) {
@@ -25,7 +26,7 @@ export class BackendAPIClientError extends Error {
 
       if (response) {
         status = response.status;
-        detail = CommonSchemas.isObjectErrorResponseSchema(response.data)
+        detail = BackendAPISchemas.isObjectErrorResponseSchema(response.data)
           ? response.data
           : {
               type: "axios_error",
@@ -39,6 +40,7 @@ export class BackendAPIClientError extends Error {
                 },
               ],
             };
+        message = detail.errors[0].detail || DEFAULT_ERROR_MESSAGE;
       }
     } else if (error instanceof Error) {
       message = error.message;
@@ -71,12 +73,33 @@ type AxiosRequestWithPayload = <T = any, R = AxiosResponse<T>, D = any>(
 
 export class BackendAPIClient {
   readonly baseURL: string;
+  protected readonly csrfCookieName: string;
   private readonly backendAPI: AxiosInstance;
 
-  constructor(baseURL: string, timeout: number) {
+  constructor(
+    baseURL: string,
+    timeout: number,
+    csrfCookieName: string = "csrftoken",
+    withCredentials: boolean = false,
+  ) {
     const headers = { "Content-Type": "application/json" };
     this.baseURL = baseURL;
-    this.backendAPI = axios.create({ baseURL, timeout, headers });
+    this.csrfCookieName = csrfCookieName;
+    this.backendAPI = axios.create({ baseURL, timeout, headers, withCredentials});
+
+    if (withCredentials) {
+      this.backendAPI.interceptors.request.use(
+        (config) => {
+          config.headers["x-csrftoken"] = this.getCSRFToken();
+          return config;
+        },
+        (error) => Promise.reject(error)
+      );
+    }
+  }
+
+  getCSRFToken(): string | undefined {
+    return getCookie(this.csrfCookieName);
   }
 
   _safe_request_without_payload(
