+++ packages/common/src/apis/index.ts
@@ -1,9 +1,9 @@
-import CommonSchemas from "../schemas";
+import BackendAPISchemas from "../schemas/backendAPI";
 import { BackendAPIClient } from "./client";
 
 namespace BackendAPIs {
-  export const listSiteMaps = (client: BackendAPIClient) => () => client.get<CommonSchemas.FlattenedSiteMapSchema[]>("v1/cms/sitemap/");
-  export const retrievePage = (client: BackendAPIClient) => (id: string) => client.get<CommonSchemas.PageSchema>(`v1/cms/page/${id}/`);
+  export const listSiteMaps = (client: BackendAPIClient) => () => client.get<BackendAPISchemas.FlattenedSiteMapSchema[]>("v1/cms/sitemap/");
+  export const retrievePage = (client: BackendAPIClient) => (id: string) => client.get<BackendAPISchemas.PageSchema>(`v1/cms/page/${id}/`);
 }
 
 export default BackendAPIs;
